+++ src/templates/webpack.config.js
@@ -8,8 +8,7 @@
  * @link http://www.ibitux.com
  */
 
-const config = require('./webpack-yii2.json');
-
+const argv = require('yargs').argv;
 const webpack = require('webpack');
 const path = require('path');
 const fs = require('fs');
@@ -17,7 +16,22 @@ const AssetsPlugin = require('assets-webpack-plugin');
 const CleanWebpackPlugin = require('clean-webpack-plugin');
 const CompressionPlugin = require('compression-webpack-plugin');
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
+
 const prodFlag = process.argv.indexOf('-p') !== -1;
+
+var confPath = './webpack-yii2.json';
+if(argv.env && argv.env.config) {
+    confPath = path.join(__dirname, argv.env.config, 'webpack-yii2.json');
+}
+if(!fs.existsSync(confPath)) {
+    throw 'Error: file "' + confPath + '" not found.';
+}
+
+var config = require(confPath);
+if (argv.env && argv.env.config) {
+    config.sourceDir = path.relative(__dirname, argv.env.config);
+}
+
 module.exports = {
     entry: config.entry,
     context: path.resolve(__dirname, config.sourceDir, config.subDirectories.sources),
@@ -76,6 +90,7 @@ module.exports = {
             }
         })
     ],
+    externals: config.externals,
     module: {
         rules: [
             {
@@ -109,7 +124,7 @@ module.exports = {
                 })
             },
             {
-                test: /\.scss$/,
+                test: /\.s[ac]ss$/,
                 use: ExtractTextPlugin.extract({
                     fallback: 'style-loader',
                     use: ['css-loader', 'sass-loader']
@@ -125,6 +140,7 @@ module.exports = {
         ]
     },
     resolve: {
+        alias: config.alias,
         extensions: ['.tsx', '.ts', '.js']
     },
     devtool: 'source-map',